moved plot and compute functions
