+++ neware_parser/DataEngine.py
@@ -103,6 +103,82 @@ class DataEngine:
             "caps": caps,
         }
 
+    @staticmethod
+    def predicted_capacity(
+        cyc_grp_dict, mode, protocols, cycle_m, cycle_v, barcode_count,
+        degradation_model, svit_and_count,
+    ) -> dict:
+        """
+        Return:
+            dict: With the following keys:
+
+                "caps" - A list of structured arrays. One structured array consists
+                    of the cycle and capacity for one protocol. There the length of
+                    the list is the number of protocols.
+        """
+
+        cycles = [x for x in np.arange(0., 6000., 20.)]
+        my_cycle = [(cyc - cycle_m) / tf.sqrt(cycle_v) for cyc in cycles]
+        caps = []
+
+        for count, protocol in enumerate(protocols):
+
+            if protocol[-1] == "dchg":
+                sign_change = -1.
+            else:
+                sign_change = +1.
+
+            if mode == "cc":
+                target_voltage = cyc_grp_dict[protocol]["avg_last_cc_voltage"]
+                target_currents = [cyc_grp_dict[protocol][Key.I_CC_AVG]]
+            elif mode == "cv":
+                target_voltage = cyc_grp_dict[protocol][Key.V_END_AVG]
+                curr_min = abs(cyc_grp_dict[protocol][Key.I_CC_AVG])
+                curr_max = abs(cyc_grp_dict[protocol]["avg_end_current"])
+
+                if curr_min == curr_max:
+                    target_currents = np.array([curr_min])
+                else:
+                    target_currents = sign_change * np.exp(
+                        np.arange(
+                            np.log(curr_min),
+                            np.log(curr_max),
+                            .05 * (np.log(curr_max) - np.log(curr_min))
+                        )
+                    )
+            else:
+                sys.exit("Unknown mode in predicted.")
+
+            test_results = test_single_voltage(
+                my_cycle,
+                target_voltage,
+                cyc_grp_dict[protocol][Key.I_CC_AVG],
+                cyc_grp_dict[protocol][Key.I_PREV_END_AVG],
+                cyc_grp_dict[protocol][Key.V_PREV_END_AVG],
+                cyc_grp_dict[protocol][Key.V_END_AVG],
+                target_currents,
+                barcode_count, degradation_model,
+                svit_and_count[Key.SVIT_GRID],
+                svit_and_count[Key.COUNT_MATRIX]
+            )
+
+            if mode == "cc":
+                pred_cap = tf.reshape(
+                    test_results["pred_cc_capacity"], shape = [-1],
+                )
+            elif mode == "cv":
+                pred_cap = test_results["pred_cv_capacity"].numpy()[:, -1]
+            else:
+                sys.exit("Unknown mode in predicted.")
+
+            caps.append(sign_change * pred_cap)
+        return {
+            "mode": mode,
+            "protocols": protocols,
+            "caps": caps,
+            "cycles": cycles,
+        }
+
 
 # TODO(harvey): duplicate function in plot.py
 def test_single_voltage(
