+++ neware_parser/plot.py
@@ -301,105 +301,6 @@ def plot_vq(plot_params, init_returns):
         plt.close(fig)
 
 
-def plot_measured(caps, protocols, mode, ax1):
-    for count, (protocol, cap) in enumerate(zip(protocols, caps)):
-        if mode == "cc":
-            cap_type = Key.Q_CC
-        elif mode == "cv":
-            cap_type = Key.Q_CV
-        else:
-            sys.exit("Unknown mode in measured.")
-
-        ax1.scatter(
-            cap[Key.N],
-            cap[cap_type],
-            c = COLORS[count],
-            s = 5,
-            label = make_legend(protocol)
-        )
-
-
-def compute_predicted(
-    cyc_grp_dict, mode, protocols, cycle_m, cycle_v, barcode_count,
-    degradation_model, svit_and_count,
-) -> dict:
-    """
-    Return:
-        dict: With the following keys:
-
-            "caps" - A list of structured arrays. One structured array consists
-                of the cycle and capacity for one protocol. There the length of
-                the list is the number of protocols.
-    """
-
-    cycles = [x for x in np.arange(0., 6000., 20.)]
-    my_cycle = [(cyc - cycle_m) / tf.sqrt(cycle_v) for cyc in cycles]
-    caps = []
-
-    for count, protocol in enumerate(protocols):
-
-        if protocol[-1] == "dchg":
-            sign_change = -1.
-        else:
-            sign_change = +1.
-
-        if mode == "cc":
-            target_voltage = cyc_grp_dict[protocol]["avg_last_cc_voltage"]
-            target_currents = [cyc_grp_dict[protocol][Key.I_CC_AVG]]
-        elif mode == "cv":
-            target_voltage = cyc_grp_dict[protocol][Key.V_END_AVG]
-            curr_min = abs(cyc_grp_dict[protocol][Key.I_CC_AVG])
-            curr_max = abs(cyc_grp_dict[protocol]["avg_end_current"])
-
-            if curr_min == curr_max:
-                target_currents = np.array([curr_min])
-            else:
-                target_currents = sign_change * np.exp(
-                    np.arange(
-                        np.log(curr_min),
-                        np.log(curr_max),
-                        .05 * (np.log(curr_max) - np.log(curr_min))
-                    )
-                )
-        else:
-            sys.exit("Unknown mode in predicted.")
-
-        test_results = test_single_voltage(
-            my_cycle,
-            target_voltage,
-            cyc_grp_dict[protocol][Key.I_CC_AVG],
-            cyc_grp_dict[protocol][Key.I_PREV_END_AVG],
-            cyc_grp_dict[protocol][Key.V_PREV_END_AVG],
-            cyc_grp_dict[protocol][Key.V_END_AVG],
-            target_currents,
-            barcode_count, degradation_model,
-            svit_and_count[Key.SVIT_GRID],
-            svit_and_count[Key.COUNT_MATRIX]
-        )
-
-        if mode == "cc":
-            pred_cap = tf.reshape(
-                test_results["pred_cc_capacity"], shape = [-1],
-            )
-        elif mode == "cv":
-            pred_cap = test_results["pred_cv_capacity"].numpy()[:, -1]
-        else:
-            sys.exit("Unknown mode in predicted.")
-
-        caps.append(sign_change * pred_cap)
-    return {
-        "mode": mode,
-        "protocols": protocols,
-        "caps": caps,
-        "cycles": cycles,
-    }
-
-
-def plot_predicted(caps, cycles, ax1):
-    for count, pred_cap in enumerate(caps):
-        ax1.plot(cycles, pred_cap, c = COLORS[count])
-
-
 def plot_capacities(
     cyc_grp_dict, cycle_m, cycle_v, barcode_count,
     degradation_model, svit_and_count, fig,
@@ -416,7 +317,7 @@ def plot_capacities(
         measured_data = DataEngine.measured_capacity(
             cyc_grp_dict, mode, protocols,
         )
-        predicted_data = compute_predicted(
+        predicted_data = DataEngine.predicted_capacity(
             cyc_grp_dict, mode, protocols, cycle_m, cycle_v,
             barcode_count, degradation_model, svit_and_count,
         )
@@ -425,11 +326,11 @@ def plot_capacities(
 
         ax1 = fig.add_subplot(6, 1, 1 + off)
 
-        plot_measured(
+        PlotEngine.measured_capacity(
             measured_data["caps"], measured_data["protocols"],
             measured_data["mode"], ax1
         )
-        plot_predicted(
+        PlotEngine.predicted_capacity(
             predicted_data["caps"], predicted_data["cycles"], ax1,
         )
